+++ src/domain/shared/components/SocialLinks/icons/LinkedIn.tsx
