+++ src/domain/shared/components/SocialLinks/models/SocialNetworks.ts
